added an if statement to check for the tslint.json file before trying to delete. If no file, skip.

+++ scripts/postinstall.ts
@@ -130,5 +130,5 @@ console.log(`INFO: Deleting tslint.info unnecessary file`);
 
 const TSLINT_FILEPATH = path.resolve(path.join(projectPath, 'tslint.json'));
 
-// delete file
-fs.unlinkSync(TSLINT_FILEPATH);
+// delete file (if it exists)
+if (fs.existsSync(TSLINT_FILEPATH)) { fs.unlinkSync(TSLINT_FILEPATH); }
